+++ demos/demo/demo.js
@@ -48,7 +48,7 @@ var demoTrack = function(name){
       // source: 'http://205.186.156.50/global-offshore/assets/audio/ocean_sounds',
       source: 'audio/' + document.getElementById(name+'-source').value,
       sourceMode: document.getElementById(name+'-source-mode').value,
-      gain: 1.0,
+      gain: 0, //1.0,
       looping: document.getElementById(name+'-looping').checked,
       muted: document.getElementById(name+'-muted').checked,
       autoplay: document.getElementById(name+'-autoplay').checked,